#14 Update plantuml and tomcat versions, switch checksum calculation
This commit updates the versions of plantuml-server and Tomcat. Since the
homepage offers PGP and SHA512 checksums the verification of the tomcat binary
changes from SHA256 to SHA512 for the ease of use (go to website, copy checksum,
use checksum). The SHA256 checksum for plantuml-server must still be calculated
manually.

This commit also re-words the single RUN statements for the Docker build stages
because these can be extended more easily than multiple calls within the same
RUN command. It should be noted that the number of container layers do not play
a role for builder stages.